Separate file perm tests
